reSWIGged
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxPython/trunk@78004 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/osx_cocoa/_gdi_wrap.cpp
@@ -2883,7 +2883,7 @@ SWIG_From_unsigned_SS_int  (unsigned int value)
 
 SWIGINTERN long wxColour_GetPixel(wxColour *self){
             #ifndef __WXGTK3__
-            return self->GetPixel();
+                return (long)self->GetPixel();
             #else
             return -1;
             #endif
